+++ src/test/kotlin/net/pwall/json/schema/codegen/CodeGeneratorEmptyClassTest.kt
@@ -1,0 +1,33 @@
+package net.pwall.json.schema.codegen
+
+import kotlin.test.Test
+import kotlin.test.expect
+
+import java.io.File
+import java.io.StringWriter
+
+class CodeGeneratorEmptyClassTest {
+
+    @Test fun `should output empty class`() {
+        val input = File("src/test/resources/test-empty-object.schema.json")
+        val codeGenerator = CodeGenerator()
+        codeGenerator.baseDirectoryName = "dummy"
+        val stringWriter = StringWriter()
+        codeGenerator.outputResolver =
+                CodeGeneratorTestUtil.outputCapture("dummy", emptyList(), "TestEmpty", "kt", stringWriter)
+        codeGenerator.basePackageName = "com.example"
+        codeGenerator.generate(input)
+        expect(expectedExample) { stringWriter.toString() }
+    }
+
+    companion object {
+
+        const val expectedExample =
+"""package com.example
+
+class TestEmpty
+"""
+
+    }
+
+}